Rename "handlers" namespace to "handler"
