+++ server/src/Handler/AbstractHandler.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Cora\Handlers;
+namespace Cora\Handler;
 
 use Psr\Http\Message\ServerRequestInterface as Request;
 use Psr\Http\Message\ResponseInterface as Response;
