+++ server/src/Handler/Feedback/CoverabilityFeedback.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace Cora\Handlers\Feedback;
+namespace Cora\Handler\Feedback;
 
 use Psr\Http\Message\ServerRequestInterface as Request;
 use Psr\Http\Message\ResponseInterface as Response;
 use Slim\Exception\HttpBadRequestException;
 
-use Cora\Handlers\AbstractHandler;
+use Cora\Handler\AbstractHandler;
 use Cora\Service\GetFeedbackService;
 use Cora\Views\Factory\ViewFactory;
 use Cora\Views\Factory\FeedbackViewFactory;
