+++ server/src/Handler/Petrinet/RegisterPetrinet.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace Cora\Handlers\Petrinet;
+namespace Cora\Handler\Petrinet;
 
 use Psr\Http\Message\ServerRequestInterface as Request;
 use Psr\Http\Message\ResponseInterface as Response;
 use Slim\Exception\HttpBadRequestException;
 
-use Cora\Handlers\AbstractHandler;
+use Cora\Handler\AbstractHandler;
 use Cora\Service\RegisterPetrinetService;
 use Cora\Views\Factory\ViewFactory;
 use Cora\Views\Factory\PetrinetCreatedViewFactory;
