+++ server/src/Handler/User/GetUser.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace Cora\Handlers\User;
+namespace Cora\Handler\User;
 
 use Psr\Http\Message\ServerRequestInterface as Request;
 use Psr\Http\Message\ResponseInterface as Response;
 use Slim\Exception\HttpBadRequestException;
 use Slim\Exception\HttpNotFoundException;
 
-use Cora\Handlers\AbstractHandler;
+use Cora\Handler\AbstractHandler;
 use Cora\Service\GetUserService;
 use Cora\Views\Factory\ViewFactory;
 use Cora\Views\Factory\UserViewFactory;
