+++ server/src/Handler/User/RegisterUser.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace Cora\Handlers\User;
+namespace Cora\Handler\User;
 
 use Psr\Http\Message\ServerRequestInterface as Request;
 use Psr\Http\Message\ResponseInterface as Response;
 use Slim\Exception\HttpBadRequestException;
 
-use Cora\Handlers\AbstractHandler;
+use Cora\Handler\AbstractHandler;
 use Cora\Service\RegisterUserService;
 use Cora\Views\Factory\ViewFactory;
 use Cora\Views\Factory\UserCreatedViewFactory;
